Trim parser-based axiom time evals
