+++ .meta/README.md
@@ -274,10 +274,6 @@ Measures whether the agent can write valid APL and PromQL. If it can't, nothing 
 
 Score 0 if no query tool calls were made at all.
 
-### axiom-time-bounds (hard gate)
-
-Fails the run if any `scripts/axiom-query` call omits an explicit wrapper time window. This is distinct from syntax validity because a query like `['dataset'] | getschema` is syntactically valid APL but still disallowed unless the tool call also carries `--since` or `--from`/`--to`.
-
 ### rca-accuracy (LLM judge)
 
 **Primary scorer for correctness.**
@@ -410,7 +406,6 @@ Avg: 7212753218.4
 ├── scorers/
 │   ├── index.ts                  # Scorer exports
 │   ├── query-validity.ts         # Syntax validity + required queries (hard gate)
-│   ├── axiom-time-bounds.ts      # Wrapper time-window enforcement for axiom-query (hard gate)
 │   ├── rca.ts                    # LLM judge (Gemini) with keyword fallback
 │   ├── evidence.ts               # Tool usage + keyword presence + data point citations
 │   └── efficiency.ts             # Budget compliance + failure rate + redundancy detection
