+++ .meta/scorers/query-error-classification.ts
@@ -26,8 +26,9 @@ const PATTERNS: Record<Exclude<QueryFailureKind, 'none'>, RegExp[]> = {
     /\bunknown flag\b/i,
     /\bmissing arguments?\b/i,
     /\bno query provided via stdin\b/i,
-    /\bexplicit _time bound\b/i,
-    /\bevery axiom apl query must include an explicit _time bound\b/i,
+    /\bexplicit time window\b/i,
+    /\bmissing time window\b/i,
+    /\buse either --since or --from\/--to\b/i,
     /\busage:\s+(?:axiom-query|grafana-query)\b/i,
   ],
   unknown_target: [/\bunknown dataset\b/i, /\bunknown metric\b/i],
