+++ .meta/toolbox/wasm/apl-parser.wasm
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:01382902abccfde61008d2aa07792d48e3bef0bbbfcd3b4725d8ecbfe7bf34f1
-size 2003953
+oid sha256:3ea05086d09e26cd99adad21441a3fdb73859299895e5ad0e392d706bc522ce1
+size 1967043
