+++ skill/scripts/init
@@ -72,7 +72,7 @@ done
 
 # Check prerequisites
 MISSING_DEPS=""
-for dep in curl jq node; do
+for dep in curl jq; do
     if ! command -v "$dep" >/dev/null 2>&1; then
         MISSING_DEPS+="$dep "
     fi
@@ -83,8 +83,8 @@ fi
 
 if [[ -n "$MISSING_DEPS" ]]; then
     echo -e "${YELLOW}Missing dependencies: ${MISSING_DEPS}${NC}"
-    echo "  macOS: brew install coreutils curl jq node"
-    echo "  Linux: apt install coreutils curl jq nodejs"
+    echo "  macOS: brew install coreutils curl jq"
+    echo "  Linux: apt install coreutils curl jq"
     echo ""
 fi
